FEAT: Convert JS in TS
