+++ src/App.tsx
@@ -1,4 +1,4 @@
-function App() {
+const App = () => {
     return <div>portfolio 3d</div>
 }
