+++ src/components/About.tsx
@@ -1,0 +1,5 @@
+const About = () => {
+    return <div>About</div>
+}
+
+export default About
