+++ src/components/Contact.tsx
@@ -1,0 +1,5 @@
+const Contact = () => {
+    return <div>Contact</div>
+}
+
+export default Contact
