--- src/components/Experience.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Experience = () => {
-  return (
-    <div>Experience</div>
-  )
-}
-
-export default Experience
