--- src/components/Feedbacks.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Feedbacks = () => {
-  return (
-    <div>Feedbacks</div>
-  )
-}
-
-export default Feedbacks
