+++ src/components/Feedbacks.tsx
@@ -1,0 +1,5 @@
+const Feedbacks = () => {
+    return <div>Feedbacks</div>
+}
+
+export default Feedbacks
