--- src/components/Hero.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Hero = () => {
-  return (
-    <div>Hero</div>
-  )
-}
-
-export default Hero
