+++ src/components/Hero.tsx
@@ -1,0 +1,5 @@
+const Hero = () => {
+    return <div>Hero</div>
+}
+
+export default Hero
