+++ src/components/Loader.tsx
@@ -1,0 +1,5 @@
+const Loader = () => {
+    return <div>Loader</div>
+}
+
+export default Loader
