+++ src/components/Navbar.tsx
@@ -1,0 +1,5 @@
+const Navbar = () => {
+    return <div>Navbar</div>
+}
+
+export default Navbar
