--- src/components/Tech.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Tech = () => {
-  return (
-    <div>Tech</div>
-  )
-}
-
-export default Tech
