+++ src/components/Tech.tsx
@@ -1,0 +1,5 @@
+const Tech = () => {
+    return <div>Tech</div>
+}
+
+export default Tech
