--- src/components/Works.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Works = () => {
-  return (
-    <div>Works</div>
-  )
-}
-
-export default Works
