+++ src/components/Works.tsx
@@ -1,0 +1,5 @@
+const Works = () => {
+    return <div>Works</div>
+}
+
+export default Works
