--- src/components/canvas/Ball.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Ball = () => {
-  return (
-    <div>Ball</div>
-  )
-}
-
-export default Ball
