+++ src/components/canvas/Ball.tsx
@@ -1,0 +1,5 @@
+const Ball = () => {
+    return <div>Ball</div>
+}
+
+export default Ball
