--- src/components/canvas/Computers.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Computers = () => {
-  return (
-    <div>Computers</div>
-  )
-}
-
-export default Computers
