+++ src/components/canvas/Computers.tsx
@@ -1,0 +1,5 @@
+const Computers = () => {
+    return <div>Computers</div>
+}
+
+export default Computers
