--- src/components/canvas/Earth.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Earth = () => {
-  return (
-    <div>Earth</div>
-  )
-}
-
-export default Earth
