+++ src/components/canvas/Earth.tsx
@@ -1,0 +1,5 @@
+const Earth = () => {
+    return <div>Earth</div>
+}
+
+export default Earth
