--- src/components/canvas/Stars.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Stars = () => {
-  return (
-    <div>Stars</div>
-  )
-}
-
-export default Stars
