+++ src/components/canvas/Stars.tsx
@@ -1,0 +1,5 @@
+const Stars = () => {
+    return <div>Stars</div>
+}
+
+export default Stars
