+++ src/components/canvas/index.ts
@@ -1,0 +1,6 @@
+import EarthCanvas from './Earth.tsx'
+import BallCanvas from './Ball.tsx'
+import ComputersCanvas from './Computers.tsx'
+import StarsCanvas from './Stars.tsx'
+
+export { EarthCanvas, BallCanvas, ComputersCanvas, StarsCanvas }
