+++ src/components/index.ts
@@ -1,0 +1,24 @@
+import { BallCanvas, ComputersCanvas, EarthCanvas, StarsCanvas } from './canvas'
+import Hero from './Hero.tsx'
+import Navbar from './Navbar.tsx'
+import About from './About.tsx'
+import Tech from './Tech.tsx'
+import Experience from './Experience.tsx'
+import Works from './Works.tsx'
+import Feedbacks from './Feedbacks.tsx'
+import Contact from './Contact.tsx'
+
+export {
+    Hero,
+    Navbar,
+    About,
+    Tech,
+    Experience,
+    Works,
+    Feedbacks,
+    Contact,
+    EarthCanvas,
+    BallCanvas,
+    ComputersCanvas,
+    StarsCanvas,
+}
